% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.dr.R
\docType{package}
\name{fastDR-package}
\alias{fastDR-package}
\title{fastDR: Fast Doubly Robust Estimation}
\description{
Uses generalized boosted models to compute non-parametric propensity score estimates and a propensity score weighted outcome regression model to compute doubly robust estimates of treatment effects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gregridgeway/fastDR}
}

}
\author{
\strong{Maintainer}: Greg Ridgeway \email{gridge@upenn.edu}

}
