% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.dr.R
\name{print.fastDR}
\alias{print.fastDR}
\title{Print a fastDR object}
\usage{
\method{print}{fastDR}(x, type = "outcome", model = "dr", ...)
}
\arguments{
\item{x}{An object of class \code{fastDR}.}

\item{type}{A string taking value \code{"outcome"} or \code{"complete"}.}

\item{model}{A string taking value \code{"un"}, \code{"ps"}, or \code{"dr"}.}

\item{...}{Arguments passed to \code{print.default}.}
}
\value{
\code{print.fastDR()} is called for its side effect of printing a summary
of treatment effect estimates. It returns \code{NULL} invisibly.
}
\description{
Display basic information about a \code{fastDR} object.
}
\details{
If \code{type = "outcome"} then \code{print.fastDR()} will output the treatment effect
estimate. Which treatment effect estimate depends on the value of \code{model}.
Presumably, \code{"dr"} will be needed, but an unadjusted and propensity score
weighted option are available. Results are reported on the response scale,
not as odds ratios or rate ratios. \code{print.fastDR()} will produce percentage
point differences for binomial outcomes, rate differences for Poisson
outcomes, and differences for all others.

If \code{type = "complete"} then \code{print.fastDR()} will print the entire effects
results.

See \code{\link[=fastDR]{fastDR()}} for an example of its use.
}
\seealso{
\code{\link[=fastDR]{fastDR()}}
}
\author{
Greg Ridgeway \email{gridge@upenn.edu}
}
\keyword{models}
