% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.dr.R
\name{fastDR}
\alias{fastDR}
\title{Fast Doubly Robust Estimation}
\usage{
fastDR(
  form.list,
  data,
  y.dist = "gaussian",
  estimand = "ATT",
  n.trees = 3000,
  interaction.depth = 3,
  shrinkage = 0.01,
  verbose = FALSE,
  ps.only = FALSE,
  keepGLM = TRUE,
  smooth.lm = 0,
  par_details = gbmParallel(1, 1024)
)
}
\arguments{
\item{form.list}{A list of formulas giving the components of the model,
outcomes, treatment indicator, covariates, observation weights, and
observation keys. Consider using \code{\link[=make.fastDR.formula]{make.fastDR.formula()}} to assist in
creating \code{form.list}.}

\item{data}{A data frame containing the variables in the model.}

\item{y.dist}{The distribution assumed for the outcome. These should be
specified as character strings, for example \code{"gaussian"},
\code{"quasibinomial"}, or \code{"quasipoisson"}. Using \code{"binomial"} will cause
inconsequential warning messages since the model uses non-integer
propensity score weights. \code{"quasibinomial"} avoids this. If \code{y.form} has
multiple outcomes then \code{y.dist} should be a vector of strings the same
length as the number of outcomes in \code{y.dist}.}

\item{estimand}{Either \code{"ATT"} (default) for the average treatment effect on
the treated or \code{"ATE"} for the average treatment effect.}

\item{n.trees}{The total number of trees to fit for the generalized boosted
model. This is equivalent to the number of iterations and the number of
basis functions in the additive expansion. The default is most likely
appropriate.}

\item{interaction.depth}{The maximum depth of variable interactions. 1
implies an additive model, 2 implies a model with up to 2-way interactions,
etc.}

\item{shrinkage}{A shrinkage parameter applied to each tree in the
generalized boosted model expansion. Also known as the learning rate or
step-size reduction. The default is most likely appropriate.}

\item{verbose}{If \code{TRUE}, \code{fastDR()} will print out progress and performance
indicators.}

\item{ps.only}{If \code{TRUE}, \code{fastDR()} will skip the DR step and just return
the propensity scores. This is useful if you need to do non-standard
changes to the regression model, like eliminating some covariates.}

\item{keepGLM}{If \code{TRUE}, keep a copy of the \code{glm} objects produced in the
outcome model step. These models can become large with multiple outcomes
and large datasets.}

\item{smooth.lm}{A numeric value for penalizing the size of the covariates in
the DR step.}

\item{par_details}{A list containing the number of threads and
\code{array_chunk_size} to be passed to \code{gbmt} for running several gbm
computations in parallel, for example gradient calculation and split
selection. Most easily set using \code{gbmParallel}.}
}
\value{
\code{fastDR()} returns a \code{fastDR} object, a list containing:
\item{ks}{The largest KS statistic in the balance table. See \code{balance.tab}
below. A measure of the quality of the propensity score fit.}
\item{ks.un}{The largest KS statistic in the balance table before propensity
score weighting. See \code{balance.tab.un} below.}
\item{balance.tab}{The balance table showing the treatment means and
propensity score weighted control means for each of the terms specified in
\code{x.form}. For categorical features the table lists each level separately.}
\item{balance.tab.un}{The balance table showing the treatment means and
unweighted control means for each of the terms specified in \code{x.form}.}
\item{w}{The propensity score weights used in the analysis. Use \code{names()} to
extract the key associated with each weight to align with the original
data. All cases in the original data might not necessarily have a
propensity score weight if they were dropped because factor levels were
outside common support. Missing, negative, or zero sampling weights cause
\code{fastDR()} to stop.}
\item{p}{The estimated propensity scores. \code{p} also maintains the \code{key} in its
names.}
\item{n1, ESS}{The number of cases in the treatment group and the effective
sample size of the control group.}
\item{glm.un, glm.ps, glm.dr}{The generalized linear model fits using only
sampling weights (\code{glm.un}), using propensity score weights (\code{glm.ps}), and
the model producing doubly robust estimates (\code{glm.dr}).}
\item{effects}{A table showing the results. \code{effects} includes results for an
unadjusted (\code{un}), propensity score adjusted (\code{ps}), and doubly robust
estimate (\code{dr}). For each, \code{effects} includes the estimates and standard
errors for E(Y1|t=1), E(Y0|t=1), their difference, and p-value. These
results are all on the scale of the response, not odds ratios or rate
ratios.}
\item{y.dist}{Contains the value of \code{y.dist} used in the call to \code{fastDR()}.}
\item{shrinkage}{The shrinkage parameter used in the propensity score model.}
}
\description{
Doubly robust treatment effect estimation with non-parametric propensity
score estimates.
}
\details{
The \code{form.list} has the following components.

\code{y.form} is a formula specifying the outcomes, for example \code{~y}. Multiple
outcomes may be listed, for example \code{~y1 + y2}; \code{fastDR()} will conduct a
separate analysis for each of them.

\code{t.form} is a formula specifying the treatment indicator, for example
\code{~treat}. The formula can include only one binary treatment coded as 0/1.

\code{x.form} is a formula specifying the potential confounding variables that
will be included in the propensity score model and in the outcome model, for
example \code{~X1 + X2 + X2}.

\code{weights.form} is an optional formula specifying observation weights, such as
sampling weights. Sampling weights cannot be missing, negative, or equal to
zero.

\code{key.form} is a formula giving observation IDs, for example \code{~caseID}. This
is required in order to make sure that after returning results, the propensity
score weights can be correctly aligned with cases by the user.

\code{fastDR()} estimates either the average treatment effect on the treated (ATT)
or the average treatment effect (ATE).

\code{fastDR()} uses a generalized boosted model to estimate a propensity score
model (McCaffrey, Ridgeway, and Morral, 2004). It uses those propensity
scores as weights in a weighted generalized linear model to produce a doubly
robust estimate of the treatment effect.

To reuse the propensity score weights in other analyses, make sure that the
code aligns the right weight with the right case. Use the key in the data and
the key attached to the weights. For example,
\code{mydata$psw <- myfastDR$w[match(mydata$observationID, names(myfastDR$w))]}.

For ATT, treated cases receive weight equal to 1 and comparison cases receive
weight p/(1-p). Control cases with factor levels that do not appear among
treated cases are dropped from the analysis. For ATE, treated cases receive
weight equal to 1/p and comparison cases receive weight 1/(1-p). Cases in
either group with factor levels that do not appear in the opposite group are
dropped from the analysis, and \code{fastDR()} issues a warning when this occurs.

The effective sample size is the number of independent cases that would yield
the same precision as the given weighted sample. It is computed as
\deqn{\frac{(\sum_i (1-t_i)w_i)^2}{\sum_i (1-t_i)w_i^2}}

When \code{smooth.lm} is greater than 0, \code{fastDR()} uses a data augmentation method
to impose a penalty on the regression coefficients in the DR step. At a
minimum, modest penalties provide numerical stability in the event of
correlation among the covariates or certain features being highly predictive
of the outcome. \code{fastDR()} does not penalize the intercept or the treatment
indicator. \code{fastDR()} uses the data augmentation approach described in
Greenland and Mansournia (2015). This approach uses ridge regression
penalties (N(0,m)) for Gaussian outcomes, log-F(m,m) prior for logistic
regression, and gamma(m,m) prior for Poisson regression. The appropriate size
of the penalty is subjective, but values between 0.1 and 4.0 are modest
penalties for general application.

For the DR estimates, \code{fastDR()} computes exact standard errors when the
prediction data used for counterfactual means has at most 5000 rows. This
corresponds to at most 2500 treated cases for ATT and at most 2500 analysis
cases for ATE. For larger analyses, \code{fastDR()} takes a random sample of 2500
cases for computing the standard errors of the estimates of E(Y0), E(Y1), and
E(Y1-Y0). Computing these standard errors requires summarizing a large
covariance matrix of predicted values, which can require a lot of memory, but
can be efficiently estimated with a sample. Note that there will be some
amount of sampling variability; in simulations, the Monte Carlo standard
errors are typically off by at most 0.1\%.
}
\examples{
# NHANES example from survey package

data(nhanes)
nhanes <- nhanes[seq_len(300), ]
# add a unique ID to each row
nhanes$observationID <- 1:nrow(nhanes)
# recode the "treatment" (male) to a 0/1 indicator
nhanes$male <- as.numeric(nhanes$RIAGENDR == 1)
# make "race" a factor with descriptive labels
nhanes$race <- factor(nhanes$race,
                      levels = c(1, 2, 3, 4),
                      labels = c("Hispanic", "non-Hispanic white",
                                 "non-Hispanic black", "other"))

dr1 <- fastDR(list(y.form = ~HI_CHOL,          # high cholesterol (over 240mg/dl)
                   t.form = ~male,             # compare males to similar females
                   x.form = ~race + agecat,    # potential confounders
                   weights.form = ~WTMEC2YR,   # sampling weights
                   key.form = ~observationID),
              data = nhanes,
              y.dist = "quasibinomial",   # outcome is 0/1
              n.trees = 14,
              interaction.depth = 1,
              shrinkage = 0.01,
              verbose = FALSE,
              smooth.lm = 0.1,            # stabilize regression coefs
              par_details = gbmParallel(1, 1024)) # just use one core

# show balance table
round(dr1$balance.tab, 3)

# compute percentage difference
print(dr1, type = "outcome", model = "dr")

}
\references{
D. McCaffrey, G. Ridgeway, and A. Morral (2004). "Propensity score
estimation with boosted regression for evaluating adolescent substance abuse
treatment." \emph{Psychological Methods} 9(4):403-425.

J.H. Friedman (2001). "Greedy Function Approximation: A Gradient Boosting
Machine." \emph{Annals of Statistics} 29(5):1189-1232.

S. Greenland and M.A. Mansournia (2015). "Penalization, bias reduction, and
default priors in logistic and related categorical and survival regressions."
\emph{Statistics in Medicine} 34:3133-3143.

S. Greenland, M.A. Mansournia, and D.G. Altman (2014). "Sparse data bias: a
problem hiding in plain sight." \emph{The BMJ} 353:i1981.

\url{https://gregridgeway.github.io/}
}
\seealso{
\code{\link[gbm3:gbmt]{gbm3::gbmt()}}, \code{\link[gbm3:gbmParallel]{gbm3::gbmParallel()}}
}
\author{
Greg Ridgeway \email{gridge@upenn.edu}
}
\keyword{models}
